+++ js/data.js
@@ -97,7 +97,7 @@ const MODS = [
 // ai: 'chaser' persigue · 'shooter' dispara a distancia · 'erratic' persigue zigzagueando · 'boss' usa patterns
 const ENEMIES = {
   rata:        { name: 'Rata',              sprite: 'rata',       hp: 14,  dmg: 6,  spd: 78,  ai: 'chaser',  size: 9 },
-  esqueleto:   { name: 'Esqueleto',         sprite: 'esqueleto',  hp: 30,  dmg: 10, spd: 52,  ai: 'chaser',  size: 11 },
+  esqueleto:   { name: 'Esqueleto',         sprite: 'esqueleto',  hp: 30,  dmg: 10, spd: 52,  ai: 'chaser',  size: 11, skel: true },
   arquero_esq: { name: 'Esqueleto arquero', sprite: 'arquero_esq',hp: 22,  dmg: 8,  spd: 44,  ai: 'shooter', size: 11, range: 150, fireCd: 1.7, projSpd: 130 },
   murcielago:  { name: 'Murciélago',        sprite: 'murcielago', hp: 12,  dmg: 5,  spd: 115, ai: 'erratic', size: 9 },
   arana:       { name: 'Araña',             sprite: 'arana',      hp: 24,  dmg: 9,  spd: 88,  ai: 'erratic', size: 10 },
@@ -129,8 +129,8 @@ const BOSSES = {
 // ---------- Zonas (el orden define la progresión de la run) ----------
 const ZONES = [
   {
-    id: 'catacumbas', name: 'Catacumbas', floors: 2,
-    palette: { floor: '#26292f', floorAlt: '#222529', wall: '#474e58', wallDark: '#343a42', accent: '#5a7d54' },
+    id: 'torre', name: 'Torre en Ruinas', floors: 2,
+    palette: { floor: '#3a3a40', floorAlt: '#343438', wall: '#5a5a62', wallDark: '#3e3e46', accent: '#6e7d6a' },
     enemies: ['rata', 'esqueleto', 'arquero_esq'],
     boss: 'bucle', density: 1.0,
   },